move ConstellationMsg to compositing
