+++ components/webdriver_server/Cargo.toml
@@ -12,6 +12,7 @@ path = "lib.rs"
 
 [dependencies]
 base64 = "0.10"
+compositing = {path = "../compositing"}
 cookie = "0.11"
 crossbeam-channel = "0.3"
 euclid = "0.20"
