+++ components/webvr/Cargo.toml
@@ -16,6 +16,7 @@ oculusvr = ['rust-webvr/oculusvr']
 
 [dependencies]
 canvas_traits = {path = "../canvas_traits"}
+compositing = {path = "../compositing"}
 crossbeam-channel = "0.3"
 euclid = "0.20"
 ipc-channel = "0.12"
@@ -23,7 +24,6 @@ log = "0.4"
 msg = {path = "../msg"}
 rust-webvr = {version = "0.16", features = ["mock", "openvr", "vrexternal"]}
 rust-webvr-api = "0.16"
-script_traits = {path = "../script_traits"}
 servo_config = {path = "../config"}
 sparkle = "0.1"
 webvr_traits = {path = "../webvr_traits" }
